ui tests moved to their own file
